+++ xknx/io/routing.py
@@ -181,7 +181,7 @@ class Routing(Interface):
     async def connect(self) -> bool:
         """Start routing."""
         self.xknx.current_address = self.individual_address
-        await self.xknx.connection_manager.connection_state_changed(
+        self.xknx.connection_manager.connection_state_changed(
             XknxConnectionState.CONNECTING, self.connection_type
         )
         try:
@@ -192,13 +192,13 @@ class Routing(Interface):
                 type(ex).__name__,
                 ex,
             )
-            await self.xknx.connection_manager.connection_state_changed(
+            self.xknx.connection_manager.connection_state_changed(
                 XknxConnectionState.DISCONNECTED
             )
             # close udp transport to prevent open file descriptors
             self.transport.stop()
             raise CommunicationError("Routing could not be started") from ex
-        await self.xknx.connection_manager.connection_state_changed(
+        self.xknx.connection_manager.connection_state_changed(
             XknxConnectionState.CONNECTED, self.connection_type
         )
         return True
@@ -206,7 +206,7 @@ class Routing(Interface):
     async def disconnect(self) -> None:
         """Stop routing."""
         self.transport.stop()
-        await self.xknx.connection_manager.connection_state_changed(
+        self.xknx.connection_manager.connection_state_changed(
             XknxConnectionState.DISCONNECTED
         )
         self._flow_control.cancel()